backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.3
cachetools==6.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
httplib2==0.31.0
idna==3.11
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
razorpay==2.0.0
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.4
stripe==14.0.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
django-storages==1.14.6
boto3==1.42.7
twilio==9.9.0
python-dotenv==1.0.1
resend==2.19.0
# Django Channels for WebSocket support
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
whitenoise==6.7.0
dj-database-url==2.3.0
